+++ muonShieldOptimization/run_fixedTarget.py
@@ -185,21 +185,22 @@ if args.AddMuonShield or args.AddHadronAbsorberOnly:
 sensPlaneHA = ROOT.exitHadronAbsorber()
 sensPlaneHA.SetEnergyCut(args.ecut*u.GeV)
 sensPlaneHA.SetVetoPointName("PlaneHA")
-sensPlaneT = ROOT.exitHadronAbsorber()
-sensPlaneT.SetEnergyCut(args.ecut*u.GeV)
-sensPlaneT.SetVetoPointName("PlaneT")
+#Additional sensitive plane around target
+#sensPlaneT = ROOT.exitHadronAbsorber()
+#sensPlaneT.SetEnergyCut(args.ecut*u.GeV)
+#sensPlaneT.SetVetoPointName("PlaneT")
 if args.storeOnlyMuons:
     sensPlaneHA.SetOnlyMuons()
-    sensPlaneT.SetOnlyMuons()
+#    sensPlaneT.SetOnlyMuons()
 if args.skipNeutrinos:
     sensPlaneHA.SkipNeutrinos()
-    sensPlaneT.SkipNeutrinos()
+#    sensPlaneT.SkipNeutrinos()
 if args.FourDP:
     sensPlaneHA.SetOpt4DP()  # in case a ntuple should be filled with pi0,etas,omega
-    sensPlaneT.SetOpt4DP()  # in case a ntuple should be filled with pi0,etas,omega
-sensPlaneT.SetZposition(ship_geo.target.length)  # if not using automatic positioning behind default magnetized hadron absorber
+#    sensPlaneT.SetOpt4DP()  # in case a ntuple should be filled with pi0,etas,omega
+#sensPlaneT.SetZposition(ship_geo.target.length)  # if not using automatic positioning behind default magnetized hadron absorber
 run.AddModule(sensPlaneHA)
-run.AddModule(sensPlaneT)
+#run.AddModule(sensPlaneT)
 
 # -----Create PrimaryGenerator--------------------------------------
 primGen = ROOT.FairPrimaryGenerator()
@@ -308,7 +309,7 @@ sTree = t.CloneTree(0)
 nEvents = 0
 for n in range(t.GetEntries()):
     rc = t.GetEvent(n)
-    if t.PlaneHAPoint.GetEntries() > 0 or t.PlaneTPoint.GetEntries()>0:
+    if t.PlaneHAPoint.GetEntries() > 0: # or t.PlaneTPoint.GetEntries()>0:
         rc = sTree.Fill()
         nEvents += 1
     #t.Clear()